+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/CreateCommand.java
@@ -9,8 +9,8 @@ import org.bukkit.entity.Player;
 
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.commands.SubCommand;
-import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.chat.Messages;
+import io.github.scroojalix.npcmanager.common.PluginUtils;
+import io.github.scroojalix.npcmanager.common.chatutils.Messages;
 
 public class CreateCommand extends SubCommand {