+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/RenameCommand.java
@@ -9,9 +9,9 @@ import org.bukkit.command.CommandSender;
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.commands.CommandUtils;
 import io.github.scroojalix.npcmanager.commands.SubCommand;
-import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.chat.Messages;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
+import io.github.scroojalix.npcmanager.common.PluginUtils;
+import io.github.scroojalix.npcmanager.common.chatutils.Messages;
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
 
 public class RenameCommand extends SubCommand {