+++ core/src/main/java/io/github/scroojalix/npcmanager/commands/subcommands/modifications/HeadRotationModification.java
@@ -7,8 +7,8 @@ import org.bukkit.command.CommandSender;
 
 import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.commands.SubCommand;
-import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
+import io.github.scroojalix.npcmanager.common.PluginUtils;
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
 
 public class HeadRotationModification extends SubCommand {