+++ core/src/main/java/io/github/scroojalix/npcmanager/common/npc/skin/SkinType.java
@@ -1,0 +1,5 @@
+package io.github.scroojalix.npcmanager.common.npc.skin;
+
+public enum SkinType {
+    PLAYER, URL;
+}