+++ core/src/main/java/io/github/scroojalix/npcmanager/nms/interfaces/INPCLoader.java
@@ -10,7 +10,7 @@ import org.bukkit.Location;
 import org.bukkit.entity.Player;
 
 import io.github.scroojalix.npcmanager.NPCMain;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
 
 public abstract class INPCLoader {