+++ v1_13_R2/src/main/java/io/github/scroojalix/npcmanager/nms/v1_13_R2/NPCLoader.java
@@ -11,10 +11,10 @@ import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 
 import io.github.scroojalix.npcmanager.NPCMain;
+import io.github.scroojalix.npcmanager.common.PluginUtils;
+import io.github.scroojalix.npcmanager.common.npc.NPCData;
+import io.github.scroojalix.npcmanager.common.npc.equipment.NPCEquipment;
 import io.github.scroojalix.npcmanager.nms.interfaces.INPCLoader;
-import io.github.scroojalix.npcmanager.utils.PluginUtils;
-import io.github.scroojalix.npcmanager.utils.npc.NPCData;
-import io.github.scroojalix.npcmanager.utils.npc.equipment.NPCEquipment;
 import net.minecraft.server.v1_13_R2.EntityArmorStand;
 import net.minecraft.server.v1_13_R2.EnumItemSlot;
 import net.minecraft.server.v1_13_R2.MathHelper;